articles/requirements.txt
confluent-kafka
prometheus_client
requests
newspaper3k